fix(coachmark): resolved circular dependency
