+++ packages/ibm-products/src/components/Coachmark/next/Coachmark/Coachmark.tsx
@@ -22,13 +22,13 @@ import React, {
 import PropTypes from 'prop-types';
 import cx from 'classnames';
 import { getDevtoolsProps } from '../../../../global/js/utils/devtools';
-import { pkg } from '../../../../settings';
+import { CoachmarkContext, blockClass } from './context';
 import CoachmarkContent, { CoachmarkContentProps } from './CoachmarkContent';
 import { NewPopoverAlignment } from '@carbon/react';
 import { useIsomorphicEffect } from '../../../../global/js/hooks';
 
 // The block part of our conventional BEM class names (blockClass__E--M).
-export const blockClass = `${pkg.prefix}--coachmark__next`;
+
 const componentName = 'Coachmark';
 
 // NOTE: the component SCSS is not imported here: it is rolled up separately.
@@ -80,28 +80,7 @@ export type CoachmarkComponent = ForwardRefExoticComponent<
 > & {
   Content: FC<CoachmarkContentProps>;
 };
-interface CoachmarkContextType {
-  onClose?: () => void;
-  open?: boolean;
-  setOpen: (value: boolean) => void;
-  align?: NewPopoverAlignment;
-  triggerRef: RefObject<HTMLElement | null>;
-  position: { x: number; y: number };
-  contentRef: HTMLElement | null;
-  setContentRef: (value: any) => void;
-  floating?: boolean;
-}
 
-export const CoachmarkContext = createContext<CoachmarkContextType>({
-  open: false,
-  setOpen: () => {},
-  align: 'bottom',
-  triggerRef: { current: null },
-  position: { x: 0, y: 0 },
-  contentRef: null,
-  setContentRef: (value: boolean) => {},
-  floating: false,
-});
 /**
  * Coachmarks are used to call out specific functionality or concepts
  * within the UI that may not be intuitive but are important for the
